Change how material index is generated as it was messing up because of wrong assumption that materials are loaded in the index order.
Fixed warning in incinerator.cpp

+++ Tools/Incinerator/Incinerator.cpp
@@ -347,7 +347,7 @@ void Incinerator::IncinerateScene(const std::string& destinationFolder, const st
 	}
 
 	Zmey::Incinerator::GLTFLoader::ParseAndIncinerate(gltfSceneData.data(),
-		gltfSceneData.size(),
+		uint32_t(gltfSceneData.size()),
 		destinationFolder,
 		contentFolder,
 		meshFiles,
